requirements.txt
aiogram==2.9.2
aiohttp==3.6.2
aioredis==1.3.1
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
hiredis==1.1.0
idna==2.10
multidict==4.7.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
tqdm==4.53.0
typing-extensions==3.7.4.3
tzlocal==2.1
yarl==1.6.3

flask~=1.1.2
requests~=2.25.1
pandas~=1.2.2